refactor(cli): extract app.go into modular files
Reduce app.go from 943 to 316 lines (66% reduction) by extracting
components into separate files following single-responsibility principle.

New files created:
- types.go: Model struct, AppState, MenuItem definitions
- handlers.go: Menu and intent input handling
- onboarding.go: Onboarding state machine and view
- accessors.go: Public accessor methods
- keys.go: Key input constants
- help.go: Help screen rendering using UIKit
- menu.go: Menu rendering with responsive columns
- registration.go: Intent registration (7 intents)
- services.go: Service initialization helpers

Additional improvements:
- Replace all lipgloss imports with UIKit primitives
- Add CenterInTerminal, JoinVertical, JoinHorizontal to primitives/text.go
- Constantize all key strings

AI-Generated-By: OpenCode (Claude Sonnet 4)
